requirements.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
Click==7.0
selenium
python-dotenv
robin_stocks
tqdm
importlib_resources
chromedriver-autoinstaller
boto3
dynamodb_json
pandas